Relocates the formatDate func to lib/dates
